Clear the buffer when changing settings that affect the buffer size, averaging, etc.
Tweak the buffering between the hardware acquisition loop and pushing data to the client. Let the push rate be variable according to how full the (relatively small) buffer has grown.

Handle the appropriate signals to shut down the application more cleanly and signal to the client(s) that it is disconnecting.